Rename DAO file
To reflect new name of feature
