+++ tests/app/dao/test_service_whitelist_dao.py
@@ -5,7 +5,7 @@ from app.models import (
     EMAIL_TYPE,
 )
 
-from app.dao.service_whitelist_dao import (
+from app.dao.service_guest_list_dao import (
     dao_fetch_service_guest_list,
     dao_add_and_commit_guest_list_contacts,
     dao_remove_service_guest_list
